almost got trueskill working

+++ app/models.py
@@ -32,10 +32,10 @@ class Trueskillrating(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     username = db.Column(db.String(64), db.ForeignKey('user.username'))
     matchtype = db.Column(db.String(64))
-    mu = db.Column(db.DECIMAL(precision=2))
-    sigma = db.Column(db.DECIMAL(precision=2))
-    rating = db.Column(db.DECIMAL(precision=2))
-    latest_delta = db.Column(db.DECIMAL(precision=2))
+    mu = db.Column(db.Float())
+    sigma = db.Column(db.Float())
+    rating = db.Column(db.Float())
+    latest_delta = db.Column(db.Float())
 
     def __repr__(self):
         return '<TrueskillRating {}>'.format(self.username)
